requirements.txt
chess
stockfish
python-lichess